move cluster-agent to subcommand
